requirements.txt
# =============================================================================
# Pipeline ML / notebooks 01–03
# =============================================================================
# NumPy fijado a la serie 1.26 porque scikit-surprise==1.1.4 se distribuye
# como wheel compilado contra NumPy 1.x. Con NumPy 2.x aparece el error
# "numpy.core.multiarray failed to import" al hacer `from surprise import ...`.
# Polars 1.9, pandas 2.2 y scikit-learn 1.5 son compatibles con NumPy 1.26.
numpy>=1.26,<2.0
polars==1.9.*
pyarrow==18.*
pandas==2.2.*
scikit-learn==1.5.*
scikit-surprise==1.1.4
psutil==6.1.*
matplotlib==3.9.*
seaborn==0.13.*
jupyter
ipykernel
tqdm
wordcloud
joblib
# Opcional — AutoML benchmark con TPE (si la instalación falla, el notebook
# 03 detecta ImportError y usa un GridSearchCV multi-algoritmo como fallback).
# auto-surprise==0.1.8

# =============================================================================
# Demo Django (CRISP-DM Fase 6 — Deployment) en `app/`
# =============================================================================
# Stack: Django + Inertia.js + React 18 + MUI 6. El frontend vive en
# `app/frontend/` y se compila con Vite (`npm run build`). django-vite
# lee `frontend/dist/.vite/manifest.json` para servir los assets
# correctos con el hash apropiado.
Django==6.0.*
django-environ==0.13.*
inertia-django==1.2.*
django-vite==3.0.*
PyMySQL=1.1.2
